[feat] Add editingMode on Main VC
- 삭제 시에 collectionView reload 문제 수정
- editingMode 시에 순서 변경 가능하도록 수정
- reorder 추가하여 순서 변경시 저장하는 order 수정하도록 함.

+++ Reminder-Clone/ViewModels/HomeListTableViewModel.swift
@@ -30,19 +30,38 @@ class HomeListTableViewModel: NSObject {
   }
 
   func deleteCategory(indexPath: IndexPath, handler: @escaping () -> Void) {
-    DispatchQueue.global().sync {
+    DispatchQueue.global().sync { [weak self] in
+      guard let category = self?.data.remove(at: indexPath.row) else { return }
       let manager = PersistentManager.shared
-      let category = data.remove(at: indexPath.row)
+
       category.tasks
         .compactMap{ $0 as? Task }
         .forEach { manager.delete($0) }
-      manager.delete(category)
 
       DispatchQueue.main.async {
         handler()
       }
-
-      manager.saveContext()
+      
+      manager.delete(category)
+      reorder()
     }
+  }
+
+  func move(from: Int, to: Int) {
+    let category = data.remove(at: from)
+    data.insert(category, at: to)
+
+    reorder()
+  }
+
+  func reorder() {
+    for (i, v) in data.enumerated() {
+      v.set(key: .order, value: i)
+    }
+    save()
+  }
+
+  func save() {
+    PersistentManager.shared.saveContext()
   }
 }
